refactor: common 폴더 안에 있는 파일 common 접두사 제거
